Loader: Use splash screens to ask for direction
If the Platform is in Setup mode, ask permission to install the binary
signature. FIXME: need to check first to see if the signature is present

If the Platform is in user mode, try to boot the binary securely and if that
fails because of access being denied, ask the user's permission to boot the
binary.

Signed-off-by: James Bottomley <JBottomley@Parallels.com>