add new security_policy
This library file has an installable security policy which overrides
the default policy in EFI_SECURITY2_ARCH_PROTOCOL with a MOK based one before
falling back to the previous policy.

The design of this is that the firstboot system can install the new security
policy (including static keys carried in its body) and then use the standard
UEFI execution mechanisms to spawn the next programs (so no need to load and
link the executables).  The security policy change remains in force until the
first boot loader exits.

This is required because the new breed of EFI bootloaders (like gummiboot) use
the standard EFI execution calls to start kernels.

Signed-off-by: James Bottomley <JBottomley@Parallels.com>