simple_file: fix missing files problem
If the filter is NULL, simple file dereferences *NULL.  It just so happens on
a lot of systems this is zero and everything works, but on sime it isn't and
thus no files appear.  Fix this by explicitly checking for NULL and replacing
it with a pointer to the empty string.

Signed-off-by: James Bottomley <JBottomley@Parallels.com>