security_policy: Consult the image table to find dbx forbidden keys
Since the basis of the ease of matching on Mok is hashes only, we can't tell
when we're executing an efi utility that is signed with a revoked key.  We can
fix this by looking at the entries in the image execution information table
and seeing why the image execution was failed.

There's a complication in this in that very few manufacturers seem to have
implemented the image execution information table to spec (most miss out the
required application name), so we have to parse it heuristically and try to
check for errors.

Signed-off-by: James Bottomley <JBottomley@Parallels.com>