efi-updatevar: add ability to update in User Mode
For this to work, you must possess the secret key for the relevant
signing variable (PK for PK and KEK or KEK for db and dbx).  We
actually check the private key against all public keys in the variable
to make sure everything is OK and then do a signed update.

so to add the hash of HelloWorld.efi to db, the command issued would be

efi-updatevar -k KEK.key -b HelloWorld.efi db

Signed-off-by: James Bottomley <JBottomley@Parallels.com>