sign-efi-sig-list: fix timestamps
The original bug which required timestamps a year in the future is long gone,
so kill the hardcoding and use the correct timestamp from the current date.

Also fix a race condition where noPK.auth could have the same timestamp as
PK.auth on fast build machines and thus fail to update.

Convert all times to platform local and use ISO date format for the -t option.

Finally zero the timestamp for update requests as required by the spec

Signed-off-by: James Bottomley <JBottomley@Parallels.com>