console.c: Fix longstanding bug (causing crashes)
The switch between a variable length selector and a fixed one is wrong.  It's
hardcoded to be rows - 10 instead of rows - 6 - title_lines.  This means when
you have exactly the right number of rows in the selector it switches over to
a variable selector too early and displays non-existent selector lines (which
may run off the end of memory and crash).

Also tidy up the selector to keep a blank line between the title and the
beginning of the selector.

Signed-off-by: James Bottomley <JBottomley@Parallels.com>